Add formik validation
